Implemented substitution cipher encryption though could not find a way to ignore other characters that didn't cause errors (like segmentation faults), but it works effectively without spaces in the message variable
Also, for a while my program has been randomly crashing when the compile button is clicked, sometimes it happens but sometimes it doesn't. Same goes for using './a.out' in the terminal which sometimes causes a seg fault. Hopefully I can clear up the issue in my lab, otherwise marking may be annoying.